fix: promise refactor
rewrite tests and implementation to fit with promises
