+++ src/cordova-media-player.js
