+++ src/cordova-media-recorder.js
