+++ test/pronunciationAnalysisSpec.js
@@ -1,0 +1,154 @@
+/* eslint-disable
+ camelcase,
+ new-cap
+ */
+
+/* global
+ afterEach,
+ beforeEach,
+ describe,
+ expect,
+ it,
+ jasmine,
+ spyOn,
+ window,
+ FormData
+ */
+
+require('jasmine-ajax');
+require('jasmine-as-promised')();
+const autobahn = require('autobahn');
+var Promise = require('es6-promise').Promise;
+const its = require('../src');
+var Q = require('q');
+
+describe('Pronunciation Analyisis Websocket API interaction test', function() {
+  beforeEach(function() {
+    jasmine.Ajax.install();
+  });
+
+  afterEach(function() {
+    jasmine.Ajax.uninstall();
+  });
+
+  it('should fail streaming when websocket connection is closed', function() {
+    var api = new its.Sdk({
+      authPrincipal: 'principal',
+      authPassword: 'secret'
+    });
+
+    // Mock the audio recorder
+    function RecorderMock() {
+      this.getAudioSpecs = function() {
+        return {
+          audioFormat: 'audio/wave',
+          audioParameters: {
+            channels: 1,
+            sampleWidth: 16,
+            sampleRate: 48000
+          }
+        };
+      };
+    }
+
+    var challenge = new its.PronunciationChallenge('fb', '4', 'foo');
+    var recorder = new RecorderMock();
+
+    var old = window.WebSocket;
+
+    window.WebSocket = jasmine.createSpy('WebSocket');
+
+    api.startStreamingPronunciationAnalysis(challenge, recorder)
+      .then(function(result){
+        fail('An error should be thrown. Got ' + result);
+      })
+      .catch(function(error) {
+        expect(error.message).toBe('WebSocket connection was not open.');
+        // Restore WebSocket
+        window.WebSocket = old;
+      });
+  });
+
+  it('should start streaming a new pronunciation analysis', function() {
+    var api = new its.Sdk({
+      wsToken: 'foo',
+      wsUrl: 'ws://foo.bar',
+      authPrincipal: 'principal',
+      authPassword: 'secret'
+    });
+
+    // Mock the audio recorder
+    function RecorderMock() {
+      this.getAudioSpecs = function() {
+        return {
+          audioFormat: 'audio/wave',
+          audioParameters: {
+            channels: 1,
+            sampleWidth: 16,
+            sampleRate: 48000
+          },
+          audioUrl: 'https://api.itslanguage.nl/download/Ysjd7bUGseu8-bsJ'
+        }
+      };
+      this.hasUserMediaApproval = function() {
+        return true;
+      };
+      this.isRecording = function() {
+        return false;
+      };
+      this.addEventListener = function(name, method) {
+        if (name === 'dataavailable') {
+          method(1);
+        }
+        method();
+      };
+      this.removeEventListener = function() {
+      };
+
+      this.hasUserMediaApproval = function() {
+        return true;
+      }
+    }
+
+    var challenge = new its.PronunciationChallenge('fb', '4', 'foo');
+    var recorder = new RecorderMock();
+    var stringDate = '2014-12-31T23:59:59Z';
+    var fakeResponse = {
+      created: new Date(stringDate),
+      updated: new Date(stringDate),
+      audioFormat: 'audio/wave',
+      audioParameters: {
+        channels: 1,
+        sampleWidth: 16,
+        sampleRate: 48000
+      },
+      audioUrl: 'https://api.itslanguage.nl/download/Ysjd7bUGseu8-bsJ'
+    };
+
+    function SessionMock() {
+      this.call = function() {
+        var d = autobahn.when.defer();
+        d.resolve(fakeResponse);
+        return d.promise;
+      };
+    }
+
+    api._session = new SessionMock();
+    api._sessionBackup = api._session.call;
+    spyOn(api._session, 'call').and.callThrough();
+    spyOn(api, '_wordsToModels').and.returnValue('OK FROM FAKE WORDSTOMODELS');
+
+    var output = api.startStreamingPronunciationAnalysis(
+      challenge, recorder);
+
+    return output
+      .then(function(result) {
+        expect(api._session.call).toHaveBeenCalled();
+        expect(api._session.call).toHaveBeenCalledWith(
+          'nl.itslanguage.pronunciation.init_analysis', [],
+          {trimStart: 0.15, trimEnd: 0.0});
+      }).catch(function(error) {
+        fail('No error should be thrown: ' + error);
+      });
+  });
+});
